+++ actions/urban_dictionary_MOD.js
